feat: combine server/island wrappers with `visleClient` prop
Move the island-vs-server decision from import resolution time to render
time. When the same component is used both with `v-client:load` and
without, the component now correctly renders as an island or plain server
component based on the `visleClient` prop generated by the directive
transform.
