+++ src/server/internal.ts
@@ -1,3 +1,2 @@
 export { islandSymbol } from './symbol.js'
-export { createServerWrapper } from './server-wrapper.js'
-export { createIslandWrapper } from './island-wrapper.js'
+export { createComponentWrapper } from './component-wrapper.js'
